api-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
httpx==0.24.1
pydantic[email]==2.6.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.7.0
redis>=6.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-cloud-pubsub
bcrypt==3.2.2
asyncpg==0.29.0
fastapi-limiter==0.2.0